backend/requirements.txt
boto3
anthropic
psycopg2-binary
redis
tavily-python
aws-xray-sdk
openai
